Move work to masking identifiers directory, set up service config
