+++ plugin-hrm-form/src/components/common/MaskingIdentifiers/EyeCloseIcon.tsx
