+++ plugin-hrm-form/src/components/common/MaskingIdentifiers/EyeOpenIcon.tsx
