+++ plugin-hrm-form/src/types/types.ts
@@ -268,7 +268,6 @@ export type FeatureFlags = {
   enable_conferencing: boolean; // Enables Conferencing UI and replaces default Twilio components and behavior
   enable_lex: boolean; // Enables consuming from Lex bots
   backend_handled_chat_janitor: boolean; // [Temporary flag until all accounts are migrated] Enables handling the janitor from taskrouter event listeners
-  enable_unmasking_masked_voice_calls: boolean; // Enables unmasking of masked voice calls
 };
 /* eslint-enable camelcase */
