+++ plugin-hrm-form/src/utils/setUpComponents.tsx
@@ -36,7 +36,7 @@ import SettingsSideLink from '../components/sideLinks/SettingsSideLink';
 import CaseListSideLink from '../components/sideLinks/CaseListSideLink';
 import StandaloneSearchSideLink from '../components/sideLinks/StandaloneSearchSideLink';
 import ManualPullButton from '../components/ManualPullButton';
-import ViewTaskNumber from '../components/common/ViewTaskNumber';
+import ViewTaskNumber from '../components/common/MaskingIdentifiers/ViewTaskNumber';
 import { AddOfflineContactButton, OfflineContactTask } from '../components/OfflineContact';
 import { chatCapacityUpdated } from '../states/configuration/actions';
 import { Box, Column, HeaderContainer, TaskCanvasOverride } from '../styles/HrmStyles';
